app/requirements.txt
requests>=2.26.0
urllib3>=1.26,<1.27
charset_normalizer>=2.0,<2.1
feedparser
chromadb
pymupdf
tqdm
sentence_splitter
tiktoken
nltk
pymupdf4llm
markdown-it-py
torch
transformers
llama_index
langchain
chromadb
huggingface_hub
ollama
langchain-ollama
langchain-community
llama-index==0.12.32
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
llama-index-llms-ollama
llama-index-llms-langchain
fastapi
uvicorn